requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
mem0ai>=1.0.0
yfinance>=0.2.40
feedparser>=6.0.0
gradio>=4.0.0
nltk>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
chromadb>=0.5.0